server/requirements.txt
fastapi[standard]
uvicorn
SQLAlchemy
psycopg2
bcrypt
PyJWT
cloudinary
python-dotenv